refactor : Dealer fight method를 Players로 옮기고, GameResults 생성 로직을 for loop 없이 만들도록 수정
