+++ src/test/kotlin/next/step/blackjack/domain/dealer/DealerTest.kt
@@ -156,73 +156,6 @@ class DealerTest : DescribeSpec({
             }
         }
 
-        context("fight") {
-            it("GameResults 제공") {
-                val players = Players.of(
-                    setOf(
-                        Player.of(
-                            PlayerName.of("unfinished17"),
-                            Cards.of(
-                                listOf(
-                                    Card.of(CardFace.SEVEN, CardSymbol.CLUB),
-                                    Card.of(CardFace.TEN, CardSymbol.HEART)
-                                )
-                            )
-                        ),
-                        Player.of(
-                            PlayerName.of("blackjack"),
-                            Cards.of(
-                                listOf(
-                                    Card.of(CardFace.ACE, CardSymbol.CLUB),
-                                    Card.of(CardFace.TEN, CardSymbol.HEART)
-                                )
-                            )
-                        ),
-                        Player.of(
-                            PlayerName.of("finished"),
-                            Cards.of(
-                                listOf(
-                                    Card.of(CardFace.TEN, CardSymbol.CLUB),
-                                    Card.of(CardFace.TEN, CardSymbol.HEART),
-                                    Card.of(CardFace.ONE, CardSymbol.HEART)
-                                )
-                            )
-                        ),
-                        Player.of(
-                            PlayerName.of("burst"),
-                            Cards.of(
-                                listOf(
-                                    Card.of(CardFace.TEN, CardSymbol.CLUB),
-                                    Card.of(CardFace.TEN, CardSymbol.HEART),
-                                    Card.of(CardFace.TEN, CardSymbol.SPADE)
-                                )
-                            )
-                        )
-                    )
-                )
-                val dealer = Dealer.of(
-                    Cards.of(
-                        listOf(
-                            Card.of(CardFace.ACE, CardSymbol.CLUB),
-                            Card.of(CardFace.TEN, CardSymbol.HEART)
-                        )
-                    )
-                )
-
-                dealer.fight(players) shouldBe GameResults(
-                    mapOf(
-                        GameResult.WIN to 3,
-                        GameResult.LOSE to 0,
-                        GameResult.TIE to 1
-                    ),
-                    mapOf(
-                        "unfinished17" to GameResult.LOSE,
-                        "blackjack" to GameResult.TIE,
-                        "finished" to GameResult.LOSE,
-                        "burst" to GameResult.LOSE,
-                    )
-                )
-            }
-        }
+       
     }
 })
